apt_udeb: allow downloading of non-udeb packages, with fallout
it becomes mandatory to call download_debs subsequent to download_udebs,
but that's fine because we have to call it with a list of deb packages
necessary to add the 'grub2' bootloader to a minimal live system.  When
running with a full system (e.g., xfce-desktop) this list is slight
overkill since for example libreetype6 is likely to be installed anyhow;
but these are small packages compared to the whole installed system so
we'll let it go.